Moved end points to folder by category
