+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.FunctionHost/Health/HealthStatus.cs
@@ -18,7 +18,7 @@ using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.Extensions.Logging;
 
-namespace GreenEnergyHub.Charges.FunctionHost
+namespace GreenEnergyHub.Charges.FunctionHost.Health
 {
     public class HealthStatus
     {
